flask/requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
google-api-core==1.16.0
google-auth==1.11.2
google-cloud==0.34.0
google-cloud-vision==0.42.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
uWSGI==2.0.18
Werkzeug==1.0.0
